_archive/model_evaluation/requirements.txt
# Model Evaluation Requirements

# Build tools
setuptools>=69.0.0
wheel>=0.42.0

# Vector database
pymilvus>=2.3.0

# Embedding models
sentence-transformers>=2.5.0
transformers>=4.36.0
# torch>=2.1.0  # Installed separately by quick_start.ps1 with CUDA support (GPU) or CPU fallback

# Azure Storage
azure-storage-blob>=12.19.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
